requirements.txt
# Web 框架
flask>=2.0.1
Werkzeug>=2.0.1
gunicorn>=20.1.0

# API 和網路
openai>=1.3.7
httpx>=0.24.0

# 音訊處理
pydub>=0.25.1
SpeechRecognition>=3.10.0

# 資料處理
numpy>=1.24.0
scipy>=1.10.0

# 環境變數和配置
python-dotenv>=0.19.0

# 開發工具
pytest>=7.0.0
black>=22.0.0
flake8>=6.0.0